change folder component refs #75
