+++ components/folder.vue
@@ -1,6 +1,6 @@
 <template>
     <div class="side-contents">
-            <h3>マイフォルダーリポジトリ</h3>
+            <h4>共有したレポート</h4>
         <div class="folder-content">
             <div class="" v-for="shareReport in shareReports" :key="shareReport.index">
                 <nuxt-link :to="{name: 'folders-id',params:{id:shareReport.shareUser + shareReport.index}}">
@@ -14,34 +14,21 @@
     </div>
 </template>
 <script>
-import { db,firebase } from '~/plugins/firebase'
 export default {
+    props:['shareReports'],
     data(){
         return{
-            shareReports:[],
-            userAddress:null
+            // shareReports:[],
+            // userAddress:null
         }
     },
-     async mounted () {
-        let accounts = await this.$web3.eth.getAccounts()
-        this.userAddress = accounts[0]
-        if(this.userAddress != null){
-            db.collection('reports').onSnapshot((snapshot)=>{
-                snapshot.docChanges().forEach((change)=>{
-                const doc = change.doc
-              if(change.type === 'added' && doc.data().shareUser == this.userAddress){
-                this.shareReports.push({id: doc.id, ...doc.data()})
-              }
-            })
-          })    
-        }
-     }
 }
 </script>
 
 <style scoped>
-h3{
+h4{
     text-align: center;
+    padding: 15px;
 }
 .folder-content{
     overflow: scroll;
@@ -52,12 +39,20 @@ h3{
     margin: 10px;
 }
 .side-contents{
-    background:#fff;
-    height: 100%;
+    /* background:#fff; */
+    position: fixed;
+    max-height: 450px;
+    min-width: 200px;
+    overflow: hidden;
+    margin: 10px;
+    margin-top: 30px;
+    /* margin-right: 50px; */
+    right: 20px;
+    padding-bottom: 50px;
+    border-radius: 10px;
+    background: #f4f4f4;
+    box-shadow:  8px 8px 25px #8b8b8b, 
+                -8px -8px 25px #ffffff;
 }
 
-.upload-demo{
-    width:auto;
-    margin: 10px;
-}
 </style>
