+++ pages/walletPage.vue
@@ -3,14 +3,11 @@
   <div class="app-layout">
     <Header/>
     <div class="main-contents">
-      <div class="side-content">
-        <Folder/>
-      </div>
       <div class="main-content">
         <div class="about-account">
           <h4>アカウント</h4>
           <el-avatar src="https://cube.elemecdn.com/0/88/03b0d39583f48206768a7534e55bcpng.png"></el-avatar>
-          <p>0x5A2B93AB2bAe9D319b49d1AeB54840f1C8D0918c</p>
+          <p>{{userAddress}}</p>
         </div>
         <h3>所持中のレポートークン:50RPT</h3>
         <div class="wallet_btn">
@@ -30,15 +27,10 @@
           <el-button type="primary" icon="el-icon-sell" class="receive_btn">取得する</el-button>
         </div>
         <div class="wallet-detail_content">
-          <el-tabs type="card">
-            <el-tab-pane label="送信履歴">
-              <p>2020/08/24 10RPT to 0x5A2B93AB2bAe9D319b49d1Adkrid40f1Cdaferfa</p>
-            </el-tab-pane>
-            <el-tab-pane label="取得履歴">
-              <p>2020/08/24 +50RPT from 0x5A2Bdkeijadofislerfjleiorfnse0f1C8D0918c</p>
-            </el-tab-pane>
-          </el-tabs>
         </div>
+      </div>
+      <div class="side-content">
+        <Folder :shareReports="shareReports"/>
       </div>
     </div>
     <Upload/>
@@ -48,6 +40,7 @@
 
 <script>
 import Header from '~/components/header.vue'
+import { db,firebase } from '~/plugins/firebase'
 
 export default {
   components: {
@@ -56,37 +49,31 @@ export default {
   data() {
     return {
       number: 0,
-      inputNumber: 0
+      shareReports:[],
+      userAddress:null
     }
   },
   methods: {
-    getNumber: async function () {
-      let ret = await this.$reportTokenContract.methods.get().call()
-      console.log(this.$reportTokenContract)
-      console.log(ret)
-      this.number = ret
-    },
-
     purchaseToken: async function () {
       let ret = await this.$reportTokenContrat.methods.purchaseToken(ownAddress, sendValue).call()
       console.log(this.$reportTokenContract)
       console.log(ret)
       this.number = ret
     },
-
-    setNumber: async function () {
-      let accounts = await this.$web3.eth.getAccounts()
-      let account = accounts[0]
-      console.log(accounts)
-      console.log(this.inputNumber)
-      let ret = await this.$reportTokenContract.methods.set(this.inputNumber).send({from: account})
-      console.log(ret)
-    },
   },
-
-  mounted() {
-    console.log('Current Block Number')
-    this.$web3.eth.getBlockNumber().then(console.log)
+  async mounted () {
+    let accounts = await this.$web3.eth.getAccounts()
+    this.userAddress = accounts[0]
+        if(this.userAddress != null){
+            db.collection('reports').onSnapshot((snapshot)=>{
+                snapshot.docChanges().forEach((change)=>{
+                const doc = change.doc
+              if(change.type === 'added' && doc.data().shareUser == this.userAddress){
+                this.shareReports.push({id: doc.id, ...doc.data()})
+              }
+            })
+          })    
+        }
   }
 
 
@@ -114,6 +101,7 @@ p {
 .side-content{
   width: 300px;
   height: 400px;
+  z-index: 1;
 }
 
 .about-account {
